src/Model_integrations/requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
numpy
pandas
joblib
psycopg2-binary
python-dotenv
